move db code to package + change track to song
