--- .sequelizerc
@@ -1,8 +1,0 @@
-const { resolve } = require('path');
-
-module.exports = {
-  config: resolve('config/database.config.js'),
-  migrationsPath: resolve('apps/web/src/db/migrations'),
-  modelsPath: resolve('apps/web/src/db/models'),
-  seedersPath: resolve('apps/web/src/db/seeders'),
-};
