+++ apps/web/package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@contentful/rich-text-react-renderer": "15.17.2",
     "@contentful/rich-text-types": "16.3.0",
+    "db": "workspace:*",
     "@emotion/cache": "11.11.0",
     "@emotion/react": "11.11.1",
     "@emotion/styled": "11.11.0",
@@ -29,15 +30,12 @@
     "graphql-request": "6.1.0",
     "lucide-react": "0.277.0",
     "mapbox-gl": "2.15.0",
-    "mysql2": "3.6.1",
     "next": "13.4.19",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-intersection-observer": "9.5.2",
     "react-is": "18.2.0",
     "react-map-gl": "7.1.6",
-    "sequelize": "6.33.0",
-    "sequelize-typescript": "2.1.5",
     "shared-core": "workspace:*",
     "sharp": "0.32.5",
     "swr": "2.2.2",
@@ -51,7 +49,6 @@
     "@types/react": "18.2.21",
     "@types/react-dom": "18.2.7",
     "eslint-config-dg": "workspace:*",
-    "sequelize-cli": "6.6.1",
     "tsconfig": "workspace:*",
     "typescript": "5.2.2"
   }
