+++ apps/web/src/components/homepage/SpotifyCard.tsx
@@ -6,7 +6,7 @@ import { TrackListing } from 'components/spotify/TrackListing';
  * Shows a card with the latest data from Spotify
  */
 export function SpotifyCard() {
-  const { data: track } = useData('latest/track');
+  const { data: track } = useData('latest/song');
   if (!track) {
     return null;
   }
