+++ packages/db/.sequelizerc
@@ -1,0 +1,8 @@
+const { resolve } = require('path');
+
+module.exports = {
+  config: resolve('database.config.js'),
+  migrationsPath: resolve('src/migrations'),
+  modelsPath: resolve('src/models'),
+  seedersPath: resolve('src/seeders'),
+};
