+++ packages/db/package.json
@@ -7,11 +7,21 @@
     "build:serve": "pscale connect dg --port 3309",
     "build:analyze": "pscale connect dg --port 3309",
     "db": "sequelize",
-    "format": "prettier --write ."
+    "format": "prettier --write .",
+    "lint": "eslint . --cache --fix",
+    "lint:types": "tsc"
+  },
+  "dependencies": {
+    "dotenv-mono": "github:dgattey/dotenv-mono#dgattey--fix-priority",
+    "mysql2": "3.6.1",
+    "sequelize": "6.33.0",
+    "sequelize-typescript": "2.1.5"
   },
   "devDependencies": {
-    "sequelize": "6.33.0",
+    "eslint-config-dg": "workspace:*",
+    "pg-hstore": "2.3.4",
     "sequelize-cli": "6.6.1",
-    "sequelize-typescript": "2.1.5"
+    "tsconfig": "workspace:*",
+    "typescript": "5.2.2"
   }
 }
