+++ packages/db/src/models/StravaDetailedActivity.ts
