+++ packages/webhooks/package.json
@@ -2,7 +2,7 @@
   "name": "webhooks",
   "private": true,
   "scripts": {
-    "dev": "cloudflared tunnel run dg",
+    "dev": "cloudflared --loglevel fatal tunnel run dg",
     "webhook": "NODE_ENV=development tsx src/webhookCommand.ts",
     "format": "prettier --write .",
     "lint": "eslint . --cache --fix",
